Revert unnecessary conversion

+++ jpo-conflictmonitor/src/main/java/us/dot/its/jpo/conflictmonitor/monitor/processors/BsmMessageCountProgressionProcessor.java
@@ -106,8 +106,8 @@ public class BsmMessageCountProgressionProcessor<Point> extends ContextualProces
                         BsmProperties previousProperties = previousState.getProperties();
                         BsmProperties currentProperties = thisState.getProperties();
 
-                        int previousMessageCount = Math.toIntExact(previousProperties.getMsgCnt());
-                        int currentMessageCount = Math.toIntExact(currentProperties.getMsgCnt());
+                        int previousMessageCount = previousProperties.getMsgCnt();
+                        int currentMessageCount = currentProperties.getMsgCnt();
 
                         if (previousHash == currentHash && previousMessageCount == currentMessageCount) {
                             // No change
@@ -135,7 +135,7 @@ public class BsmMessageCountProgressionProcessor<Point> extends ContextualProces
         bsmData.getProperties().setOdeReceivedAt(null);
 
         // Save original msgCnt value and set them to 0
-        final int originalMsgCnt = Math.toIntExact(bsmData.getProperties().getMsgCnt());
+        final int originalMsgCnt = bsmData.getProperties().getMsgCnt();
         bsmData.getProperties().setMsgCnt(0);
 
         int hash = bsmData.hashCode();
@@ -153,10 +153,10 @@ public class BsmMessageCountProgressionProcessor<Point> extends ContextualProces
 
         BsmMessageCountProgressionEvent event = new BsmMessageCountProgressionEvent();
         event.setMessageType("BSM");
-        event.setMessageCountA(Math.toIntExact(previousProperties.getMsgCnt()));
+        event.setMessageCountA(previousProperties.getMsgCnt());
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
         event.setTimestampA(previousState.getProperties().getTimeStamp().format(formatter));
-        event.setMessageCountB(Math.toIntExact(currentProperties.getMsgCnt()));
+        event.setMessageCountB(currentProperties.getMsgCnt());
         event.setTimestampB(thisState.getProperties().getTimeStamp().format(formatter));
         if (thisState.getProperties().getId() != null) {
             event.setVehicleId(thisState.getProperties().getId());
